requirements.txt
instructor==1.3.1
ipython==8.12.3
langchain==0.2.1
langchain_community==0.2.1
openai==1.30.3
pydantic==2.7.1
python-dotenv==1.0.1
boto3
pgvector
psycopg2-binary
feedparser
bs4
